Initial commit for lockrun recipe.
Lockrun is a small executable intended to protect jobs in the crontab
from overrunning. This recipe includes the latest version of the lockrun
executable, builds it, and installs it to the specified install
prefix. There is also an LWRP designed to wrap the cron resource so you
can set it and forget it.